requirements.txt
line-bot-sdk
flask
pyrebase
paho-mqtt